refactor: 플레이리스트 이전 서비스 경로 Text DesignSystem으로 분리
